Update source to be service_endpoint_url in CSW Form
The Service endpoint is commonly the only information
an analyst has readily available. As a result it
makes more sense to have it be the first piece of
information that a user enters.

In addition this helps pave the way for future functionality
that will involve pulling metadata from the service
endpoint to help the users automatically fill out the
CSW form and reduce the risk of formatting errors.