Adjust capabilites view for mobile (#629)
The geonode_anywhere app was not being utilized and the capabilities view was only checking to see if a Oauth application was created with a client_id of 'anywhere' and the geonode_anywhere app was installed. After speaking with the anywhere team it was identified that they were not using any `/anywhere` endpoints. 

The view now checks to see if Oauth application was created with a client_id of 'anywhere' and that Django setting ANYWHERE_ENABLED is set to True.